feat: Add Basket Add Items API with JSON responses
This commit serves to add two things. One, it adds an IAP app in ecommerce extensions to isolate and distinguish any code related/needed in fulfilling In-app purchases on the ecommerce side.
Two, it adds the first API needed for the  In-App purchase flow - the basket add items API. This API is almost a copy of the original BasketAddItemsView (https://github.com/edx/ecommerce/blob/17d18356218a5a5c4c4cc82a1967ddc8c555f962/ecommerce/extensions/basket/views.py#L421)
except that instead of redirecting the user to the Checkout MFE or basket page, it returns JSON responses to make it suited for mobile API calls. In addition, it removes the voucher/coupon/discount steps
as currently they are not the part of In-App purchases MVP. The new API endpoint is essentially doing all the things that the original endpoint would do minus the discount flow.
This is done _only_ to make the API compatible with the mobile API calls. The expected parameters are same to the original endpoint.  The expected JSON responses will be defined up next alongwith the status codes that would help determine the success/failure of the items addition to the user's basket.

GET /api/iap/v1/basket/add/?sku=123

Expected Parameters: sku (productId on the mobile side)

Responses:
- Already purchased error => "error": "You have already purchased these products", status code: 422

- BadRequestException => "error": <exception's message that raised this exception>, status code: 404

- Success => "success": "Course added to the basket successfully", basket_id: <basket id of the basket the course was added to>, status code: 200

The endpoint like the original endpoint depends on the prepare_basket util method to handle the basket addition. So, any change in the said util method could potentially affect the working of this as well as the original endpoint.

Ticket: https://openedx.atlassian.net/browse/LEARNER-8392

EPIC: LEARNER-8386 (https://openedx.atlassian.net/browse/LEARNER-8386)